refactor: rename `toId` to `getPoolId`
