+++ packages/v4-sdk/src/utils/getPoolId.test.ts
@@ -1,8 +1,8 @@
 import { zeroAddress } from 'viem'
 import { expect, test } from 'vitest'
-import { PoolKey, toId } from './poolKey'
+import { PoolKey, getPoolId } from './getPoolId'
 
-test('BinPoolKey toId', () => {
+test('BinPoolKey getPoolId', () => {
   const binPoolKey: PoolKey<'Bin'> = {
     currency0: zeroAddress,
     currency1: zeroAddress,
@@ -12,10 +12,10 @@ test('BinPoolKey toId', () => {
       binStep: 10,
     },
   }
-  expect(toId(binPoolKey)).toBe('0xed39e13743a8db9e0fab025534f857fc13f5af4aace6d3deb9907b0bb9af7063')
+  expect(getPoolId(binPoolKey)).toBe('0xed39e13743a8db9e0fab025534f857fc13f5af4aace6d3deb9907b0bb9af7063')
 })
 
-test('CLPoolKey toId', () => {
+test('CLPoolKey getPoolId', () => {
   const clPoolKey: PoolKey<'CL'> = {
     currency0: zeroAddress,
     currency1: zeroAddress,
@@ -28,5 +28,5 @@ test('CLPoolKey toId', () => {
       },
     },
   }
-  expect(toId(clPoolKey)).toBe('0x0f2e9d5df51e26258038163179bbedcf8f4318a4a456e61934f3aab1a504db58')
+  expect(getPoolId(clPoolKey)).toBe('0x0f2e9d5df51e26258038163179bbedcf8f4318a4a456e61934f3aab1a504db58')
 })
